# api/doc/openapi.yaml
openapi: 3.1.0
info:
  title: Vigilo Framework API
  description: |
    REST API for Vigilo Framework: user accounts, organizations, locations,
    developer teams, and lambda configuration / deployment.

    ## Authentication
    Every user endpoint except `POST /vf/users/authenticate` and the signed report-download
    route requires an `API_KEY`
    request header. The key
    identifies the calling user; the server resolves the caller and enforces
    per-object access (e.g. a user may only read/update their own profile; only an
    admin may create users, and only within their own organization).

    Callers authenticate and obtain an `API_KEY` by posting their credentials to
    `POST /vf/users/authenticate`, which returns the user profile and a signed,
    time-limited key.

    Lambda execution, current-state, variable, and alert endpoints instead require the
    time-limited lambda JWT supplied to the running lambda in the `LAMBDA_API_KEY`
    header. A lambda may start only its current run, update only its own assignment
    variables, its assigned location's shared variables, its assigned location,
    and devices actively assigned to that location. Alert creation is restricted
    to that same location and its actively assigned devices.

    ## Responses
    Except for binary/JSON value and signed report-download routes, every `/vf/*` endpoint returns
    **HTTP 200** with an `application/json` body for both success and failure.
...
    HTTP status:

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /vf/locations:
    get:
      tags:
      - Locations
      summary: List organization locations
      description: Administrators see all locations in their organization. Developers see only testing locations
        granted to their active teams.
      operationId: listLocations
      responses:
        '200':
          description: Visible locations envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationsResponse'
    post:
      tags:
      - Locations
      summary: Create a location
      description: Organization administrator only. Creates an OPERATIONAL or TESTING location. Location type
        cannot subsequently be changed.
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '200':
          $ref: '#/components/responses/LocationResult'
  /vf/locations/{locationId}:
    get:
      tags:
      - Locations
      summary: Get a location
      description: Administrators may inspect organization locations. Developers may inspect granted testing
        locations. Returns current state and resident records, which confer no login access.
      operationId: getLocation
      parameters:
      - $ref: '#/components/parameters/LocationIdPath'
      responses:
        '200':
          $ref: '#/components/responses/LocationDetailResult'
    put:
      tags:
      - Locations
      summary: Update a location
      description: Admin-only. Replaces the location name and address; identity, organization, and lifecycle
        dates are preserved.
      operationId: updateLocation
      parameters:
      - $ref: '#/components/parameters/LocationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          $ref: '#/components/responses/LocationResult'
    delete:
      tags:
      - Locations
      summary: Delete a location
      description: Organization administrator only. Soft deletion is blocked by active resident or device assignments.
      operationId: deleteLocation
      parameters:
      - $ref: '#/components/parameters/LocationIdPath'
      responses:
        '200':
          description: Empty success envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
  /vf/locations/{locationId}/current-state:
    put:
      tags: [Lambda Runtime State]
      summary: Update a location's current state
      description: |
        Stores the current state for the lambda's assigned location. The server sets
        `stateDate` to the receipt time. A lambda cannot update another location.
      operationId: updateLocationCurrentState
      security:
        - LambdaKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/LocationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateLocationCurrentStateRequest' }
      responses:
        '200':
          description: Updated location current-state envelope (check resultCode).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LocationCurrentStateResponse' }
...
  /vf/locations/{locationId}/devices/{deviceUuid}:
    post:
      tags:
      - Devices
      summary: Assign a device to a location
      description: Organization administrator only. Device and location must belong to the caller organization
        and their test/operational designation must match. Device must have no other active assignment.
      operationId: assignDeviceToLocation
      parameters:
      - $ref: '#/components/parameters/LocationIdPath'
      - $ref: '#/components/parameters/DeviceUuidPath'
      responses:
        '200':
          $ref: '#/components/responses/DeviceLocationAssignmentResult'
    delete:
      tags:
      - Devices
      summary: Cancel a device location assignment
      description: Admin-only. Closes the active assignment and retains its history.
      operationId: cancelDeviceLocationAssignment
      parameters:
      - $ref: '#/components/parameters/LocationIdPath'
      - $ref: '#/components/parameters/DeviceUuidPath'
      responses:
        '200':
          description: Empty success envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
...
  /vf/teams:
    get:
      tags:
      - Dev Teams
      summary: List developer teams
      description: Administrators list teams within their organization, optionally filtered by member userId.
        Developers list their own active teams only.
      operationId: listDevTeams
      parameters:
      - name: userId
        in: query
        required: false
        description: Restrict results to teams this user is a member of.
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Teams envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevTeamsResponse'
    post:
      tags:
      - Dev Teams
      summary: Create a developer team
      description: Organization administrator only. Designates a developer account in the same organization as
        owner, adds the owner as a member, and atomically creates and grants a default testing location.
      operationId: createDevTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDevTeamRequest'
      responses:
        '200':
          description: Created team envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevTeamResponse'
  /vf/teams/{devTeamId}:
    get:
      tags:
      - Dev Teams
      summary: Get a developer team
      operationId: getDevTeam
      parameters:
      - name: devTeamId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Team envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevTeamResponse'
      description: Accessible to the team organization administrator and active team members only.
  /vf/teams/{devTeamId}/users/{userId}:
    post:
      tags:
      - Dev Teams
      summary: Add a developer team member
      description: Organization administrator only. Members must be developer login accounts in the team organization.
        Transfer ownership before removing an owner; removed members immediately lose inherited testing access.
      operationId: addDevTeamMember
      parameters:
      - $ref: '#/components/parameters/DevTeamIdPath'
      - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Empty success envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
    delete:
      tags:
      - Dev Teams
      summary: Remove a developer team member
      description: Organization administrator only. Members must be developer login accounts in the team organization.
        Transfer ownership before removing an owner; removed members immediately lose inherited testing access.
      operationId: deleteDevTeamMember
      parameters:
      - $ref: '#/components/parameters/DevTeamIdPath'
      - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Empty success envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
...
  /vf/locations/{locationId}/residents/{residentId}:
    post:
      tags:
      - Locations
      summary: Assign a resident to a location
      description: Organization administrator only. Resident and location must belong to the caller organization.
        Real residents can only be assigned to operational locations; synthetic residents to testing locations.
        A resident has at most one active assignment. Cancellation retains history.
      operationId: assignResidentToLocation
      parameters:
      - $ref: '#/components/parameters/LocationIdPath'
      - name: residentId
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Created assignment envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentLocationResponse'
    delete:
      tags:
      - Locations
      summary: Cancel a resident's location assignment
      description: Organization administrator only. Resident and location must belong to the caller organization.
        Real residents can only be assigned to operational locations; synthetic residents to testing locations.
        A resident has at most one active assignment. Cancellation retains history.
      operationId: cancelResidentLocationAssignment
      parameters:
      - $ref: '#/components/parameters/LocationIdPath'
      - name: residentId
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Empty success envelope (check resultCode).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
  /vf/residents:
    get:
      tags:
      - Residents
      summary: List organization residents (administrator only)
      security:
      - ApiKeyAuth: []
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentResponse'
    post:
      tags:
      - Residents
      summary: Register a resident without credentials (administrator only)
      security:
      - ApiKeyAuth: []
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResidentRequest'
  /vf/residents/{residentId}:
    get:
      tags:
      - Residents
      summary: Get a resident (administrator only)
      security:
      - ApiKeyAuth: []
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentResponse'
      parameters:
      - name: residentId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    put:
      tags:
      - Residents
      summary: Update resident profile (administrator only)
      security:
      - ApiKeyAuth: []
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentResponse'
      parameters:
      - name: residentId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResidentRequest'
  /vf/teams/{devTeamId}/owner/{userId}:
    put:
      tags:
      - Developer teams
      summary: Transfer team ownership to an active developer member (organization administrator only)
      security:
      - ApiKeyAuth: []
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
      parameters:
      - name: devTeamId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
  /vf/teams/{devTeamId}/testing-locations/{locationId}:
    put:
      tags:
      - Developer teams
      summary: Grant or revoke team access to a testing location in the same organization (administrator only)
      security:
      - ApiKeyAuth: []
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
      parameters:
      - name: devTeamId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
      - name: locationId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    delete:
      tags:
      - Developer teams
      summary: Grant or revoke team access to a testing location in the same organization (administrator only)
      security:
      - ApiKeyAuth: []
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
      parameters:
      - name: devTeamId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
      - name: locationId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: API_KEY
      description: Per-user API key identifying the caller.
    LambdaKeyAuth:
      type: apiKey
      in: header
      name: LAMBDA_API_KEY
      description: Time-limited JWT identifying the running lambda assignment.
  parameters:
    UserIdPath:
      name: userId
      in: path
      required: true
      schema: { type: integer, format: int32 }
...
    LocationIdPath:
      name: locationId
      in: path
      required: true
      schema: { type: integer, format: int32 }
    DeviceUuidPath:
      name: deviceUuid
      in: path
      required: true
      schema: { type: string }
...
    DevTeamIdPath:
      name: devTeamId
      in: path
      required: true
      schema: { type: integer, format: int32 }
...
  responses:
...
    LocationResult:
      description: |
        Location envelope, always HTTP 200. On success `resultCode` is 0 and
        `location` is populated; on failure `resultCode` is non-zero.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/LocationResponse' }
    LocationDetailResult:
      description: |
        Location-detail envelope, always HTTP 200. On success `resultCode` is 0,
        `location` is populated with its current state when available, and `users`
        contains all actively assigned residents.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/LocationDetailResponse' }
...
    DeviceLocationAssignmentResult:
      description: |
        Device-location assignment envelope, always HTTP 200. On success
        `assignment` contains the newly opened assignment.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DeviceLocationAssignmentResponse' }
...
  schemas:
...
    ActionResponse:
      type: object
      description: Common response envelope returned by every endpoint (always HTTP 200).
      properties:
        resultCode:
          type: integer
          format: int32
          description: 0 = success; non-zero is an ApiResultCodes value.
        resultCodeMessage:
          type: string
          nullable: true
        resultCodeDesc:
          type: string
          readOnly: true
          description: Human-readable description of resultCode.
        collectionTotalSize:
          type: integer
          format: int32
...
    ApiAddress:
      type: object
      properties:
        addrStreet1: { type: string, nullable: true }
        addrStreet2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        countryCode: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        timezone: { type: string, nullable: true }
...
    LocationsResponse:
      allOf:
        - $ref: '#/components/schemas/ActionResponse'
        - type: object
          properties:
            locations:
              type: array
              items: { $ref: '#/components/schemas/ApiLocation' }
    LocationResponse:
      allOf:
        - $ref: '#/components/schemas/ActionResponse'
        - type: object
          properties:
            location: { $ref: '#/components/schemas/ApiLocation' }
    LocationDetailResponse:
      allOf:
      - $ref: '#/components/schemas/ActionResponse'
      - type: object
        properties:
          location:
            $ref: '#/components/schemas/ApiLocationDetails'
          residents:
            type: array
            items:
              $ref: '#/components/schemas/Resident'
    ApiLocation:
      type: object
      properties:
        locationId:
          type: integer
          format: int32
        locationName:
          type: string
        createdAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          description: Any non-null value means the location is permanently deleted.
        organizationId:
          type: integer
          format: int32
        address:
          $ref: '#/components/schemas/ApiAddress'
        locationType:
          type: string
          enum:
          - OPERATIONAL
          - TESTING
          default: OPERATIONAL
          description: Immutable environment type. Testing grants never grant operational access.
...
    ApiLocationCurrentState:
      type: object
      required: [state, stateDate]
      properties:
        state: { type: string }
        stateDate: { type: string, format: date-time }
    UpdateLocationCurrentStateRequest:
      type: object
      required: [state]
      properties:
        state:
          type: string
          minLength: 1
          maxLength: 50
    LocationCurrentStateResponse:
      allOf:
        - $ref: '#/components/schemas/ActionResponse'
        - type: object
          properties:
            currentState: { $ref: '#/components/schemas/ApiLocationCurrentState' }
...
    DeviceLocationAssignmentResponse:
      allOf:
        - $ref: '#/components/schemas/ActionResponse'
        - type: object
          properties:
            assignment: { $ref: '#/components/schemas/ApiDeviceLocationAssignment' }
...
    ApiDeviceLocationAssignment:
      allOf:
        - $ref: '#/components/schemas/ApiDeviceLocation'
        - type: object
          properties:
            deviceUuid: { type: string }
...
    CreateLocationRequest:
      type: object
      required:
      - locationName
      - organizationId
      - address
      properties:
        locationName:
          type: string
        organizationId:
          type: integer
          format: int32
          minimum: 1
        address:
          $ref: '#/components/schemas/ApiAddress'
        locationType:
          type: string
          enum:
          - OPERATIONAL
          - TESTING
          default: OPERATIONAL
          description: Immutable environment type. Testing grants never grant operational access.
    UpdateLocationRequest:
      type: object
      required: [locationName, address]
      properties:
        locationName: { type: string }
        address: { $ref: '#/components/schemas/ApiAddress' }
...
    DevTeamsResponse:
      allOf:
        - $ref: '#/components/schemas/ActionResponse'
        - type: object
          properties:
            teams:
              type: array
              items: { $ref: '#/components/schemas/ApiDevTeamSummary' }
    DevTeamResponse:
      allOf:
        - $ref: '#/components/schemas/ActionResponse'
        - type: object
          properties:
            team: { $ref: '#/components/schemas/ApiDevTeamDetails' }
    CreateDevTeamRequest:
      type: object
      required:
      - name
      - ownerUserId
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
          nullable: true
        ownerUserId:
          type: integer
          minimum: 1
    ApiDevTeamSummary:
      type: object
      description: Team as returned by the list endpoint — scalar fields and member/lambda counts, without the
        nested collections.
      properties:
        devTeamId:
          type: integer
          format: int32
        ownerUserId:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
          nullable: true
        membersCount:
          type: integer
          format: int32
        lambdasCount:
          type: integer
          format: int32
        organizationId:
          type: integer
          format: int32
        testingLocationId:
          type: integer
          format: int32
    ApiDevTeamDetails:
      type: object
      description: Team as returned by the get-by-id endpoint — scalar fields plus members, lambdas, and all currently granted testing locations.
      properties:
        devTeamId:
          type: integer
          format: int32
        ownerUserId:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
          nullable: true
        members:
          type: array
          items:
            $ref: '#/components/schemas/ApiDevTeamMember'
        lambdas:
          type: array
          items:
            $ref: '#/components/schemas/ApiLambda'
        organizationId:
          type: integer
          format: int32
        testingLocations:
          type: array
          description: All granted, non-deleted testing locations, including the default only while its grant remains active. Empty when no grants remain.
          items:
            $ref: '#/components/schemas/ApiLocation'
        testingLocationId:
          description: Provisioned default location identifier; not evidence of a current access grant.
          type: integer
          format: int32
...
    ApiDevTeamMember:
      type: object
      properties:
        userId: { type: integer, format: int32 }
        startDate: { type: string, format: date-time, nullable: true }
        endDate: { type: string, format: date-time, nullable: true }
...
    ResidentLocationResponse:
      allOf:
      - $ref: '#/components/schemas/ActionResponse'
      - type: object
        properties:
          assignment:
            $ref: '#/components/schemas/ApiResidentLocation'
    ApiResidentLocation:
      type: object
      properties:
        residentId:
          type: integer
          format: int32
        locationId:
          type: integer
          format: int32
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
...
    UpdateResidentRequest:
      type: object
      required:
      - firstName
      - lastName
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
    CreateResidentRequest:
      allOf:
      - $ref: '#/components/schemas/UpdateResidentRequest'
      - type: object
        properties:
          synthetic:
            type: boolean
            default: false
            description: Immutable; synthetic residents can only be assigned to testing locations.
    ResidentResponse:
      allOf:
      - $ref: '#/components/schemas/ActionResponse'
      - type: object
        properties:
          resident:
            $ref: '#/components/schemas/Resident'
          residents:
            type: array
            items:
              $ref: '#/components/schemas/Resident'